apps/backend/pyproject.toml
[project]
name = "ledgersg-api"
version = "1.0.0"
description = "Enterprise-grade accounting API for Singapore SMBs"
requires-python = ">=3.12"
readme = "README.md"
license = {text = "AGPL-3.0"}
authors = [
    {name = "LedgerSG Team", email = "dev@ledgersg.sg"}
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: Django :: 6.0",
]

dependencies = [
# Core Framework
"django==6.0.2",
"djangorestframework==3.16.1",
"djangorestframework-simplejwt==5.5.1",
"django-cors-headers==4.9.0",
"django-filter==25.2",
"django-csp==4.0", # SEC-003: Content Security Policy

    # Database
    "psycopg[binary]==3.3.3",          # Psycopg 3 (not psycopg2)

    # Async / Task Queue
    "celery[redis]==5.6.2",
    # "django-celery-beat>=2.6,<3.0",
    "redis==6.4.0",

    # Decimal / Financial
    "py-moneyed==3.0",               # Currency-aware Decimal

    # Serialization / Validation
    "pydantic==2.12.5",                  # Internal validation

    # PDF Generation
    "weasyprint==68.1",

    # XML (Peppol PINT-SG)
    "lxml==6.0.2",

    # Utilities
    "python-decouple==3.8",
    "whitenoise==6.11.0",
    "gunicorn==25.1.0",
    "structlog==25.5.0",                      # Structured logging
    "sentry-sdk[django]==2.53.0",
    "argon2-cffi==25.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-django==4.12.0",
    "pytest-cov==7.0.0",
    "pytest-xdist==3.8.0",
    "model-bakery==1.23.3",                   # Test fixture factories
    "factory-boy==3.3.3",
    "faker==40.5.1",
    "httpx==0.28.1",                          # Async test client
    "ruff==0.15.2",                            # Linter + formatter
    "mypy==1.19.1",
    "django-stubs==5.2.9",
    "djangorestframework-stubs==3.16.8",
    "pre-commit==4.5.1",
    "ipython==9.10.0",
    "django-debug-toolbar==6.2.0",
    "django-extensions==4.1",
]

[project.scripts]
ledgersg = "manage:main"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "SIM", "TCH"]
ignore = ["S101"]  # Allow assert in tests

[tool.ruff.lint.isort]
known-first-party = ["apps", "common", "config"]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
strict = true
warn_return_any = true
warn_unused_configs = true
show_error_codes = true

[tool.django-stubs]
django_settings_module = "config.settings.development"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.testing"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers --tb=short -q"
markers = [
    "slow: marks tests as slow",
    "integration: marks integration tests",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
